Refactor AnnonymousUser to use concrete Account type
Replaced the use of the IAccount interface with the concrete
Account class in the AnnonymousUser class. Updated the `_account`
field and `Account` property to use the Account type. Added an
explicit interface implementation for `IAnnonymousUser.Account`
to provide custom behavior, including updating the `DateModified`
property when the `Account` is set.